.claude/skills/resume-scorer/requirements.txt
anthropic>=0.18.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
